+++ lib/sql_query.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 require 'erb'
+require_relative 'sql_query/config'
+require_relative 'sql_query/comment_remover'
 
 class SqlQuery
   attr_reader :connection
@@ -34,7 +36,7 @@ class SqlQuery
   end
 
   def sql
-    @sql ||= prepare_query(false)
+    @sql ||= apply_comment_removal(prepare_query(false), for_logs: false)
   end
 
   def pretty_sql
@@ -46,7 +48,7 @@ class SqlQuery
   end
 
   def prepared_for_logs
-    @prepared_for_logs ||= prepare_query(true)
+    @prepared_for_logs ||= apply_comment_removal(prepare_query(true), for_logs: true)
   end
 
   def partial(partial_name, partial_options = {})
@@ -65,15 +67,6 @@ class SqlQuery
     yield(config)
   end
 
-  class Config
-    attr_accessor :path, :adapter
-
-    def initialize
-      @path = '/app/sql_queries'
-      @adapter = ActiveRecord::Base
-    end
-  end
-
   private
 
   def prepare_query(for_logs)
@@ -89,6 +82,10 @@ class SqlQuery
       parts = file.rpartition('/')
       [parts.first, parts.last]
     end
+  end
+
+  def apply_comment_removal(sql, for_logs:)
+    CommentRemover.new(self.class.config).remove(sql, for_logs: for_logs)
   end
 
   def pretty(value)